+++ src/client/tent.c
@@ -504,7 +504,7 @@ static void CL_AddExplosions(void)
         }
 
         frac = (cl.time - ex->start) * BASE_1_FRAMETIME;
-        f = floor(frac);
+        f = floorf(frac);
 
         ent = &ex->ent;
 
@@ -1175,8 +1175,8 @@ static void CL_RailSpiral(void)
         VectorClear(p->accel);
 
         d = i * 0.1f;
-        c = cos(d);
-        s = sin(d);
+        c = cosf(d);
+        s = sinf(d);
 
         VectorScale(right, c, dir);
         VectorMA(dir, s, up, dir);
@@ -1210,9 +1210,9 @@ static void CL_RailTrail(void)
 
 static void dirtoangles(vec3_t angles)
 {
-    angles[0] = RAD2DEG(acos(te.dir[2]));
+    angles[0] = RAD2DEG(acosf(te.dir[2]));
     if (te.dir[0])
-        angles[1] = RAD2DEG(atan2(te.dir[1], te.dir[0]));
+        angles[1] = RAD2DEG(atan2f(te.dir[1], te.dir[0]));
     else if (te.dir[1] > 0)
         angles[1] = 90;
     else if (te.dir[1] < 0)
